set up `MockProvider` and `mount` helper
