+++ packages/client/package.json
@@ -48,6 +48,7 @@
     "electron >= 4"
   ],
   "devDependencies": {
+    "@apollo/react-testing": "^0.1.0-beta.3",
     "@graphql-codegen/add": "^1.2.1",
     "@graphql-codegen/cli": "^1.2.1",
     "@graphql-codegen/typescript": "^1.2.1",
