--- packages/client/src/testHelpers.ts
@@ -1,17 +1,0 @@
-import * as enzyme from "enzyme"
-import * as React from "react"
-
-let mounts: enzyme.ReactWrapper[] = []
-
-afterEach(() => {
-  for (const m of mounts) {
-    m.unmount()
-  }
-  mounts = []
-})
-
-export function mount(element: React.ReactElement): enzyme.ReactWrapper {
-  const wrapper = enzyme.mount(element)
-  mounts.push(wrapper)
-  return wrapper
-}
